requirements.txt
# Core Data Handling
pandas
numpy

# Database Connectivity (PostgreSQL)
SQLAlchemy
psycopg2-binary

# API Framework
fastapi
uvicorn[standard]

# Data Validation
pydantic

#Faker for generating fake data
faker

# Testing Framework
pytest

# Best Practice: Environment Variables (Crucial for storing database credentials securely)
python-dotenv